CHAD-17163: lazy loading of matter-window-covering sub-drivers
