+++ drivers/SmartThings/matter-window-covering/src/init.lua
@@ -1,16 +1,6 @@
--- Copyright 2022 SmartThings
---
--- Licensed under the Apache License, Version 2.0 (the "License");
--- you may not use this file except in compliance with the License.
--- You may obtain a copy of the License at
---
---     http://www.apache.org/licenses/LICENSE-2.0
---
--- Unless required by applicable law or agreed to in writing, software
--- distributed under the License is distributed on an "AS IS" BASIS,
--- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
--- See the License for the specific language governing permissions and
--- limitations under the License.
+-- Copyright 2022 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
 
 --Note: Currently only support for window shades with the PositionallyAware Feature
 --Note: No support for setting device into calibration mode, it must be done manually
@@ -376,10 +366,7 @@ local matter_driver_template = {
     capabilities.battery,
     capabilities.batteryLevel,
   },
-  sub_drivers = {
-    -- for devices sending a position update while device is in motion
-    require("matter-window-covering-position-updates-while-moving")
-  }
+  sub_drivers = require("sub_drivers"),
 }
 
 local matter_driver = MatterDriver("matter-window-covering", matter_driver_template)
