+++ drivers/SmartThings/matter-window-covering/src/matter-window-covering-position-updates-while-moving/fingerprints.lua
@@ -1,0 +1,8 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local SUB_WINDOW_COVERING_VID_PID = {
+  {0x10e1, 0x1005} -- VDA
+}
+
+return SUB_WINDOW_COVERING_VID_PID
